more test cleanup
